Moved DuplicateCourseSelectionController
