+++ app/views/candidate_interface/course_choices/duplicate_course_selection/new.html.erb
